+++ tests/unit_tests/service/ingest/test_ingest_service.py
@@ -14,6 +14,7 @@ def test_ingest_when_ok(
     family_repo_mock,
     document_repo_mock,
     event_repo_mock,
+    validation_service_mock,
 ):
     test_data = {
         "collections": [
@@ -40,7 +41,6 @@ def test_ingest_when_ok(
                 "family_import_id": "test.new.family.0",
                 "variant_name": "Original Language",
                 "metadata": {"color": ["blue"]},
-                "events": [],
                 "title": "",
                 "source_url": None,
                 "user_language_name": "",
@@ -57,12 +57,12 @@ def test_ingest_when_ok(
         ],
     }
 
-    assert ingest_service.import_data(test_data, "test") == {
+    assert {
         "collections": ["test.new.collection.0"],
         "families": ["created"],
         "documents": ["test.new.doc.0"],
         "events": ["test.new.event.0"],
-    }
+    } == ingest_service.import_data(test_data, "test")
 
 
 def test_ingest_when_db_error(corpus_repo_mock, collection_repo_mock):
@@ -80,122 +80,33 @@ def test_ingest_when_db_error(corpus_repo_mock, collection_repo_mock):
 
     with pytest.raises(RepositoryError) as e:
         ingest_service.import_data(test_data, "test")
-    assert e.value.message == "bad collection repo"
-
-
-def test_save_collections_when_import_id_wrong_format(corpus_repo_mock):
-    invalid_import_id = "invalid"
-    test_data = [
-        {
-            "import_id": invalid_import_id,
-            "title": "Test title",
-            "description": "Test description",
-        },
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_collections(test_data, "test")
-    assert e.value.message == "The import id invalid is invalid!"
-
-
-def test_save_families_when_import_id_wrong_format(
-    corpus_repo_mock, geography_repo_mock, db_client_metadata_mock
-):
-    invalid_import_id = "invalid"
-    test_data = [
-        {
-            "import_id": invalid_import_id,
-            "title": "Test",
-            "summary": "Test",
-            "geography": "Test",
-            "category": "UNFCCC",
-            "metadata": {"color": ["blue"], "size": [""]},
-            "collections": [],
-        },
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_families(test_data, "test")
-    assert e.value.message == "The import id invalid is invalid!"
+    assert "bad collection repo" == e.value.message
 
 
 def test_save_families_when_corpus_invalid(corpus_repo_mock):
     corpus_repo_mock.valid = False
 
-    test_data = [
-        {
-            "import_id": "test.new.family.0",
-            "title": "Test",
-            "summary": "Test",
-            "geography": "Test",
-            "category": "Test",
-            "metadata": {},
-            "collections": [],
-        },
-    ]
+    test_data = [{"import_id": "test.new.family.0"}]
 
     with pytest.raises(ValidationError) as e:
         ingest_service.save_families(test_data, "test")
-    assert e.value.message == "Corpus 'test' not found"
+    assert "Corpus 'test' not found" == e.value.message
 
 
-def test_save_families_when_geography_invalid(
+def test_save_families_when_data_invalid(
     corpus_repo_mock, db_client_metadata_mock, geography_repo_mock
 ):
     geography_repo_mock.error = True
     test_data = [
         {
             "import_id": "test.new.family.0",
-            "title": "Test",
-            "summary": "Test",
-            "geography": "Invalid",
-            "category": "UNFCCC",
-            "metadata": {"color": ["blue"], "size": [""]},
-            "collections": [],
+            "category": "Invalid",
         },
     ]
 
     with pytest.raises(ValidationError) as e:
         ingest_service.save_families(test_data, "test")
-    assert e.value.message == "The geography value Invalid is invalid!"
-
-
-def test_save_families_when_category_invalid(corpus_repo_mock, geography_repo_mock):
-    test_data = [
-        {
-            "import_id": "test.new.family.0",
-            "title": "Test",
-            "summary": "Test",
-            "geography": "Test",
-            "category": "Test",
-            "metadata": {},
-            "collections": [],
-        },
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_families(test_data, "test")
-    assert e.value.message == "Test is not a valid FamilyCategory"
-
-
-def test_save_families_when_collection_ids_invalid(
-    corpus_repo_mock, geography_repo_mock
-):
-    test_data = [
-        {
-            "import_id": "test.new.family.0",
-            "title": "Test",
-            "summary": "Test",
-            "geography": "Test",
-            "category": "UNFCCC",
-            "metadata": {},
-            "collections": ["invalid"],
-        },
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_families(test_data, "test")
-    assert e.value.message == "The import ids are invalid: ['invalid']"
+    assert "Invalid is not a valid FamilyCategory" == e.value.message
 
 
 # TODO: Uncomment when implementing feature/pdct-1402-validate-collection-exists-before-creating-family
@@ -223,96 +134,13 @@ def test_save_families_when_collection_ids_invalid(
 #     assert e.value.message == expected_msg
 
 
-def test_save_families_when_metadata_not_found(
-    corpus_repo_mock, geography_repo_mock, collection_repo_mock, db_client_metadata_mock
-):
-    db_client_metadata_mock.bad_taxonomy = True
-
-    test_data = [
-        {
-            "import_id": "test.new.family.0",
-            "title": "Test",
-            "summary": "Test",
-            "geography": "Test",
-            "category": "UNFCCC",
-            "metadata": {},
-            "collections": ["id.does.not.exist"],
-        },
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_families(test_data, "test")
-    assert e.value.message == "No taxonomy found for corpus"
-
-
-def test_save_families_when_metadata_invalid(
-    corpus_repo_mock, geography_repo_mock, collection_repo_mock, db_client_metadata_mock
-):
-    test_data = [
-        {
-            "import_id": "test.new.family.0",
-            "title": "Test",
-            "summary": "Test",
-            "geography": "Test",
-            "category": "UNFCCC",
-            "metadata": {},
-            "collections": ["id.does.not.exist"],
-        },
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_families(test_data, "test")
-    assert "Metadata validation failed: Missing metadata keys:" in e.value.message
-
-
-def test_save_documents_when_import_id_wrong_format():
-    invalid_import_id = "invalid"
-    test_data = [
-        {
-            "import_id": invalid_import_id,
-            "family_import_id": "test.new.family.0",
-            "variant_name": "Original Language",
-            "metadata": {"color": ["blue"]},
-            "events": [],
-            "title": "",
-            "source_url": None,
-            "user_language_name": "",
-        }
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_documents(test_data, "test")
-    assert e.value.message == f"The import id {invalid_import_id} is invalid!"
-
-
-def test_save_documents_when_family_import_id_wrong_format():
-    invalid_family_import_id = "invalid"
-    test_data = [
-        {
-            "import_id": "test.new.document.0",
-            "family_import_id": invalid_family_import_id,
-            "variant_name": "Original Language",
-            "metadata": {"color": ["blue"]},
-            "events": [],
-            "title": "",
-            "source_url": None,
-            "user_language_name": "",
-        }
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_documents(test_data, "test")
-    assert e.value.message == f"The import id {invalid_family_import_id} is invalid!"
-
-
-def test_save_documents_when_variant_empty():
+def test_save_documents_when_data_invalid():
     test_data = [
         {
             "import_id": "test.new.document.0",
             "family_import_id": "test.new.family.0",
             "variant_name": "",
             "metadata": {},
-            "events": [],
             "title": "",
             "user_language_name": "",
         },
@@ -320,25 +148,7 @@ def test_save_documents_when_variant_empty():
 
     with pytest.raises(ValidationError) as e:
         ingest_service.save_documents(test_data, "test")
-    assert e.value.message == "Variant name is empty"
-
-
-def test_save_documents_when_metadata_invalid(db_client_metadata_mock):
-    test_data = [
-        {
-            "import_id": "test.new.document.0",
-            "family_import_id": "test.new.family.0",
-            "variant_name": None,
-            "metadata": {},
-            "events": [],
-            "title": "",
-            "user_language_name": "",
-        },
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_documents(test_data, "test")
-    assert "Metadata validation failed: Missing metadata keys:" in e.value.message
+    assert "Variant name is empty" == e.value.message
 
 
 def test_validate_entity_relationships_when_no_family_matching_document():
@@ -351,7 +161,7 @@ def test_validate_entity_relationships_when_no_family_matching_document():
 
     with pytest.raises(ValidationError) as e:
         ingest_service.validate_entity_relationships(test_data)
-    assert e.value.message == f"No family with id ['{fam_import_id}'] found"
+    assert f"No family with id ['{fam_import_id}'] found" == e.value.message
 
 
 def test_save_documents_when_no_family():
@@ -364,38 +174,24 @@ def test_save_documents_when_no_family():
 
     with pytest.raises(ValidationError) as e:
         ingest_service.import_data(test_data, "test")
-    assert e.value.message == f"No family with id ['{fam_import_id}'] found"
+    assert f"No family with id ['{fam_import_id}'] found" == e.value.message
 
 
-def test_save_events_when_import_id_wrong_format():
-    invalid_import_id = "invalid"
-    test_data = [
-        {
-            "import_id": invalid_import_id,
-            "family_import_id": "test.new.family.0",
-            "event_title": "Test",
-            "date": datetime.now(),
-            "event_type_value": "Amended",
-        }
-    ]
-
-    with pytest.raises(ValidationError) as e:
-        ingest_service.save_events(test_data, "test")
-    assert e.value.message == f"The import id {invalid_import_id} is invalid!"
-
-
-def test_save_events_when_family_import_id_wrong_format():
-    invalid_import_id = "invalid"
+def test_save_events_when_data_invalid(db_client_metadata_mock):
+    invalid_event_type = "invalid"
     test_data = [
         {
             "import_id": "test.new.event.0",
-            "family_import_id": invalid_import_id,
+            "family_import_id": "test.new.family.0",
             "event_title": "Test",
             "date": datetime.now(),
-            "event_type_value": "Amended",
+            "event_type_value": invalid_event_type,
         }
     ]
 
     with pytest.raises(ValidationError) as e:
         ingest_service.save_events(test_data, "test")
-    assert e.value.message == f"The import id {invalid_import_id} is invalid!"
+    assert (
+        f"No allowed event types found for event {test_data[0]['import_id']}"
+        == e.value.message
+    )
